OCR.find_all_in_string improvements
Higher coverage, lower precision.

This is a more general algorithm that doesn't need to explicit account for specific characters that could appear in the middle of an OCR number (e.g. semicolons, full stops, spaces).

We simply look at all digits in the message as one long string of digits, then identify all OCR-valid substrings within the allowed length range.

This will generate more candidates than the old algorithm, some of which will be wrong, but the user is expected to set reasonable length conditions and to check results against actual invoices.